+++ src/cmd/update.rs
@@ -96,8 +96,6 @@ fn do_update(pack: &Package) -> Result<()> {
     let path = pack.path();
     if !path.is_dir() {
         Err(Error::plugin_not_installed(&pack.name))
-    } else if pack.local {
-        Err(Error::SkipLocal)
     } else {
         git::update(&pack.name, &path)
     }